Move event processors into proc directories
