+++ include/faint/proc/PreSelection.h
@@ -3,7 +3,7 @@
 
 #include "ROOT/RVec.hxx"
 
-#include <faint/EventProcessor.h>
+#include <faint/proc/EventProcessor.h>
 #include <faint/Types.h>
 
 namespace faint {
