+++ include/faint/proc/Weighter.h
@@ -4,7 +4,7 @@
 #include <cmath>
 #include <nlohmann/json.hpp>
 
-#include <faint/EventProcessor.h>
+#include <faint/proc/EventProcessor.h>
 #include <faint/Types.h>
 
 namespace faint {
