+++ src/Samples.cc
@@ -7,7 +7,7 @@
 #include <nlohmann/json.hpp>
 
 #include "faint/Log.h"
-#include "faint/MuonSelector.h"
+#include "faint/proc/MuonSelector.h"
 
 namespace faint {
 namespace {
