+++ src/proc/EventProcessor.cc
@@ -1,4 +1,4 @@
-#include "faint/EventProcessor.h"
+#include "faint/proc/EventProcessor.h"
 
 namespace faint {
