+++ src/proc/PreSelection.cc
@@ -1,4 +1,4 @@
-#include "faint/PreSelection.h"
+#include "faint/proc/PreSelection.h"
 
 #include "faint/Selections.h"
